error handling code cleanup and fixes
in some cases, use of x_ or xx_ can be error-prone,
due to the way $@ is handled; commands requiring
quotes, or with funny file names as arguments such
as spaces in the file name, or other special
characters, can make the x/xx functions break.

in those cases, where x/xx must not be used, the
commands use || err instead

in other cases, use of x/xx is superfluous, and has
been removed in some commands.

+++ include/option.sh
@@ -67,8 +67,7 @@ handle_coreboot_utils()
 			x_ mkdir -p "cbutils/${1}" && \
 			x_ cp "src/coreboot/${1}/util/${util}/${util}" \
 			    "cbutils/${1}"
-		[ -z "${mode}" ] || \
-			x_ rm -Rf "cbutils/${1}"
+		[ -z "${mode}" ] || x_ rm -Rf "cbutils/${1}"
 	done
 }
